pyproject.toml
[tool.poetry]
name = "car"
version = "0.1.0"
description = ""
authors = ["Umut Can <n.umut1499@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.8.2"
requests = "^2.32.3"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
psycopg2-binary = "^2.9.9"
pendulum = "^3.0.0"
dynaconf = "^3.2.5"
pandas = "^2.2.2"
jupyter = "^1.1.1"
redis = "^5.0.8"
pyjwt = "^2.9.0"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
